fix TC plugin update

+++ platform/util/src/com/intellij/openapi/util/io/FileUtil.java
@@ -357,13 +357,13 @@ public class FileUtil {
     return createTempFile(dir, prefix, suffix, create, true);
   }
 
-  public static File createTempFile(@NonNls final File dir, @NotNull @NonNls String prefix, @NonNls String suffix, final boolean create, boolean removeOnExit) throws IOException {
+  public static File createTempFile(@NonNls final File dir, @NotNull @NonNls String prefix, @NonNls String suffix, final boolean create, boolean deleteOnExit) throws IOException {
     File file = doCreateTempFile(prefix, suffix, dir);
     file.delete();
     if (create) {
       file.createNewFile();
     }
-    if (removeOnExit) {
+    if (deleteOnExit) {
       file.deleteOnExit();
     }
     return file;
@@ -371,10 +371,17 @@ public class FileUtil {
 
   @NotNull
   public static File createTempFile(@NotNull @NonNls String prefix, @NonNls String suffix) throws IOException{
+    return createTempFile(prefix, suffix, false); //false until TeamCity fixes its plugin
+  }
+
+  @NotNull
+  public static File createTempFile(@NotNull @NonNls String prefix, @NonNls String suffix, boolean deleteOnExit) throws IOException {
     File file = doCreateTempFile(prefix, suffix);
     file.delete();
     file.createNewFile();
-    file.deleteOnExit();
+    if (deleteOnExit) {
+      file.deleteOnExit();
+    }
     return file;
   }
